теперь при редактировании опроса с одинаковыми ответами выдает ошибку

+++ base/api/serializers.py
@@ -46,9 +46,13 @@ class PollSerializer(serializers.ModelSerializer):
         if 'type_voting' in validated_data:
             instance.type_voting = validated_data.get('type_voting')
         if 'choices' in self.initial_data:
-            instance.choice_set.all().delete()
-            for choice in self.initial_data.get('choices'):
-                instance.choice_set.create(choice=choice)
+            choices = self.initial_data['choices']
+            if len(choices) == len(set(choices)):
+                instance.choice_set.all().delete()
+                for choice in self.initial_data.get('choices'):
+                    instance.choice_set.create(choice=choice)
+            else:
+                raise serializers.ValidationError('Choices must be unique')
         instance.redacted_at = datetime.now()
         instance.save()
         return instance
